Add anchors which can deactivate the resizing of a decorator
Each decorator can have a vertical and a horizontal anchor. If they are
set to true, they make sure the decorator isn't resized in either the
horizontal or vertical direction (width and height).

This can be used to create decorators that can't be resized at all (e.g.
buttons in a header) or just into a certain direction.