Fix #4 - Delegate function calls to the child if necessary
This will make sure that decorators who add new functions to the public
interface of the component won't be "hidden" by decorators on top of
them.

If the called function doesn't exist in the current decorator it simply
delegates the call to its child (if it has one) through the __index
metamethod.